# stsiv-api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "media_ml",
        "version": "0.1.0"
    },
    "paths": {
        "/api/v1/user/list": {
            "get": {
                "summary": "Get Users",
                "description": "_summary_\n\nArgs:\n    user_service (UserController, optional): _description_. \n    Defaults to Depends(Factory().get_user_controller).\n\nReturns:\n    list[UserResponse]: _description_",
                "operationId": "get_users_api_v1_user_list_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/UserResponse"
                                    },
                                    "type": "array",
                                    "title": "Response Get Users Api V1 User List Get"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/v1/user/sign-up": {
            "post": {
                "summary": "Register User",
                "description": "_summary_\n\nArgs:\n    register_user_request (RegisterUserRequest): _description_\n    user_service (UserService, optional): _description_. Defaults to Depends(BaseContainer().get_user_service).\n\nReturns:\n    UserResponse: _description_",
                "operationId": "register_user_api_v1_user_sign_up_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterUserRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/user/sign-in": {
            "post": {
                "summary": "Login User",
                "description": "_summary_\n\nArgs:\n    login_user_request (LoginUserRequest): _description_\n    user_service (UserService, optional): _description_. Defaults to Depends(BaseContainer().get_user_service).\n\nReturns:\n    TokenSchema: _description_",
                "operationId": "login_user_api_v1_user_sign_in_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginUserRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/user/me": {
            "get": {
                "summary": "Get User",
                "description": "_summary_\n\nArgs:\n    current_user (CurrentUser, optional): _description_. Defaults to Depends(get_auth_user).\n    user (User, optional): _description_. Defaults to Depends(get_current_user).\n\nReturns:\n    UserResponse: _description_",
                "operationId": "get_user_api_v1_user_me_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "LoginUserRequest": {
                "properties": {
                    "token": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "email"
                            },
                            {
                                "type": "string"
                            }
                        ],
                        "title": "Token"
                    },
                    "password": {
                        "type": "string",
                        "title": "Password"
                    }
                },
                "type": "object",
                "required": [
                    "token",
                    "password"
                ],
                "title": "LoginUserRequest"
            },
...
            "RegisterUserRequest": {
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email"
                    },
                    "password2": {
                        "type": "string",
                        "maxLength": 64,
                        "minLength": 8,
                        "title": "Password2"
                    },
                    "password1": {
                        "type": "string",
                        "maxLength": 64,
                        "minLength": 8,
                        "title": "Password1"
                    },
                    "username": {
                        "type": "string",
                        "maxLength": 64,
                        "minLength": 3,
                        "title": "Username"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "surname": {
                        "type": "string",
                        "title": "Surname"
                    }
                },
                "type": "object",
                "required": [
                    "email",
                    "password2",
                    "password1",
                    "username",
                    "name",
                    "surname"
                ],
                "title": "RegisterUserRequest"
            },
            "TokenSchema": {
                "properties": {
                    "token": {
                        "type": "string",
                        "title": "Token",
                        "description": "Token"
                    },
                    "refresh_token": {
                        "type": "string",
                        "title": "Refresh Token",
                        "description": "Refresh token"
                    }
                },
                "type": "object",
                "required": [
                    "token",
                    "refresh_token"
                ],
                "title": "TokenSchema"
            },
            "UserResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid4",
                        "title": "Id",
                        "example": "a3b8f042-1e16-4f0a-a8f0-421e16df0a2f"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email",
                        "example": "john.doe@example.com"
                    },
                    "username": {
                        "type": "string",
                        "title": "Username",
                        "example": "john.doe"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "email",
                    "username"
                ],
                "title": "UserResponse",
                "description": "_summary_\n\nArgs:\n    BaseModel (_type_): _description_"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}